+++ CHANGELOG.rst
@@ -2,6 +2,15 @@
 Changelog for package pylon_ros2_camera
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.0.5 (2024-05-02)
+-------------------
+* Use acquisition timestamp if corresponding chunk is enabled.
+* Minor fix: warn on unsupported startup profile.
+
+2.0.4 (2024-02-12)
+-------------------
+* Fix access to chunk data: line status all, counter value, frame counter, exposure, timestamp. Adding dedicated test script.
+
 2.0.3 (2023-12-01)
 -------------------
 * Fix PylonROS2CameraNode plugin-component mismatch